Pull out all reusable methods Add TableResultsContext
