+++ src/components/custom/edit/dataset/AncestorIds.jsx
@@ -17,7 +17,7 @@ import Modal from 'react-bootstrap/Modal';
 import {PlusLg, QuestionCircleFill} from "react-bootstrap-icons";
 import {valid_dataset_ancestor_config} from "../../../../config/config";
 import Facets from "search-ui/components/core/Facets";
-import {TableResults} from '../../TableResults'
+import {TableResultsEntities} from '../../TableResultsEntities'
 import AncestorsTable from "./AncestorsTable";
 import CustomClearSearchBox from "../../layout/CustomClearSearchBox";
 import addons from "../../js/addons/addons";
@@ -160,7 +160,7 @@ export default class AncestorIds extends React.Component {
 
 
                                                     {wasSearched && <Results filters={filters}
-                                                                             view={TableResults} inModal={true} onRowClicked={this.changeAncestor}
+                                                                             view={TableResultsEntities} inModal={true} onRowClicked={this.changeAncestor}
                                                     />}
                                                 </div>
                                             }
