+++ src/components/custom/entities/collection/Datasets.jsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import DataTable from 'react-data-table-component'
-import {TableResults} from "../../TableResults";
+import {TableResultsEntities} from "../../TableResultsEntities";
 import SenNetAccordion from "../../layout/SenNetAccordion";
 
 function Datasets({ data }) {
 
     const getColumns = () => {
-        const {datasetColumns, defaultColumns} = TableResults({forData: true, rowFn: (row) => row ? row : null})
+        const {datasetColumns, defaultColumns} = TableResultsEntities({forData: true, rowFn: (row) => row ? row : null})
         return defaultColumns({hasMultipleEntityTypes: false, columns: datasetColumns, _isLoggedIn: true})
     }
