+++ src/config/constants.js
@@ -2,6 +2,7 @@
 export const APP_ROUTES = {
     "home": "/",
     "search": "/search",
+    "searchFiles": "/search/files",
     "login": "/login",
     "logout": "/logout",
     "notFound": "/404"
